Fill GodotcordLobbyManager with content
