+++ godotcord.cpp
@@ -23,15 +23,10 @@ void Godotcord::_bind_methods() {
     ClassDB::bind_method(D_METHOD("init", "client_id"), &Godotcord::init);
 	ClassDB::bind_method(D_METHOD("init_debug", "client_id", "instance_id"), &Godotcord::init_debug);
 	ClassDB::bind_method(D_METHOD("run_callbacks"), &Godotcord::run_callbacks);
-	ClassDB::bind_method(D_METHOD("search_lobbies", "search_parameters", "limit"), &Godotcord::search_lobbies);
-	ClassDB::bind_method(D_METHOD("get_lobbies", "limit"), &Godotcord::get_lobbies);
 
 	ClassDB::bind_method(D_METHOD("get_current_username"), &Godotcord::get_current_username);
 	ClassDB::bind_method(D_METHOD("get_current_user_discriminator"), &Godotcord::get_current_user_discriminator);
 	ClassDB::bind_method(D_METHOD("get_current_user_id"), &Godotcord::get_current_user_id);
-
-	ClassDB::bind_method(D_METHOD("set_lobby_metadata", "lobby_id", "key", "value"), &Godotcord::set_lobby_metadata);
-	ClassDB::bind_method(D_METHOD("get_lobby_metadata", "lobby_id", "key"), &Godotcord::get_lobby_metadata);
 
 	ClassDB::bind_method(D_METHOD("filter_relationships", "object", "function_name"), &Godotcord::filter_relationships);
 
@@ -41,21 +36,6 @@ void Godotcord::_bind_methods() {
 
 	ADD_SIGNAL(MethodInfo("search_result", PropertyInfo(Variant::ARRAY, "result")));
 	ADD_SIGNAL(MethodInfo("relationship_update", PropertyInfo(Variant::DICTIONARY, "relationship")));
-
-	BIND_ENUM_CONSTANT(LOCAL);
-	BIND_ENUM_CONSTANT(DEFAULT);
-	BIND_ENUM_CONSTANT(EXTENDED);
-	BIND_ENUM_CONSTANT(LOCAL);
-
-	BIND_ENUM_CONSTANT(LESS_THAN_OR_EQUAL);
-	BIND_ENUM_CONSTANT(LESS_THAN);
-	BIND_ENUM_CONSTANT(EQUAL);
-	BIND_ENUM_CONSTANT(GREATER_THAN);
-	BIND_ENUM_CONSTANT(GREATER_THAN_OR_EQUAL);
-	BIND_ENUM_CONSTANT(NOT_EQUAL);
-
-	BIND_ENUM_CONSTANT(STRING);
-	BIND_ENUM_CONSTANT(INT);
 }
 
 Error Godotcord::init(discord::ClientId clientId) {
@@ -183,173 +163,6 @@ int64_t Godotcord::get_current_user_id() {
 	discord::Result result = _core->UserManager().GetCurrentUser(&user);
 	ERR_FAIL_COND_V(result != discord::Result::Ok, 0);
 	return user.GetId();
-}
-
-void Godotcord::set_lobby_metadata(int64_t lobby_id, String key, String value) {
-
-	discord::LobbyTransaction txn{};
-	_core->LobbyManager().GetLobbyUpdateTransaction(lobby_id, &txn);
-
-	txn.SetMetadata(key.utf8(), value.utf8());
-
-	_core->LobbyManager().UpdateLobby(lobby_id, txn, [](discord::Result result) {
-		ERR_FAIL_COND(result != discord::Result::Ok);
-	});
-}
-
-String Godotcord::get_lobby_metadata(int64_t lobby_id, String key) {
-	char value[4096];
-
-	discord::Result result = _core->LobbyManager().GetLobbyMetadataValue(lobby_id, key.utf8(), value);
-
-	ERR_FAIL_COND_V(result != discord::Result::Ok, "");
-
-	return String(value);
-}
-
-void Godotcord::search_lobbies(Variant ref_params, int limit) {
-	ERR_FAIL_COND(!ref_params.is_array());
-	Array params = ref_params;
-	discord::LobbySearchQuery query;
-	discord::Result result = _core->LobbyManager().GetSearchQuery(&query);
-	ERR_FAIL_COND(result != discord::Result::Ok);
-
-	for (int i = 0; i < params.size(); i++) {
-		Variant element = params[i];
-		ERR_CONTINUE(element.get_type() != Variant::DICTIONARY);
-		Dictionary d = element;
-
-		String property = d.get("property", "");
-		int comparison = d.get("comparison", -1);
-		int cast = d.get("cast", -1);
-		Variant value = d.get("value", "");
-
-		query.Limit(limit);
-
-		if (property == "distance") {
-			ERR_CONTINUE(value.get_type() != Variant::INT);
-			discord::LobbySearchDistance distance;
-			int i_value = value;
-			switch (i_value) {
-				case LOCAL:
-					distance = discord::LobbySearchDistance::Local;
-					break;
-				case DEFAULT:
-					distance = discord::LobbySearchDistance::Default;
-					break;
-				case EXTENDED:
-					distance = discord::LobbySearchDistance::Extended;
-					break;
-				case GLOBAL:
-					distance = discord::LobbySearchDistance::Global;
-					break;
-				default:
-					continue;
-			}
-
-			query.Distance(distance);
-		}
-		else {
-			ERR_CONTINUE(property == "" || comparison == -1 || cast == -1 || value.get_type() != Variant::STRING);
-			String s_value = value;
-
-			discord::LobbySearchCast d_cast;
-			discord::LobbySearchComparison d_comp;
-
-			switch (cast) {
-				case STRING:
-					d_cast = discord::LobbySearchCast::String;
-					break;
-				case INT:
-					d_cast = discord::LobbySearchCast::Number;
-					break;
-				default:
-					continue;
-			}
-
-			switch (comparison) {
-				case LESS_THAN_OR_EQUAL:
-					d_comp = discord::LobbySearchComparison::LessThanOrEqual;
-					break;
-				case LESS_THAN:
-					d_comp = discord::LobbySearchComparison::LessThan;
-					break;
-				case EQUAL:
-					d_comp = discord::LobbySearchComparison::Equal;
-					break;
-				case GREATER_THAN:
-					d_comp = discord::LobbySearchComparison::GreaterThan;
-					break;
-				case GREATER_THAN_OR_EQUAL:
-					d_comp = discord::LobbySearchComparison::GreaterThanOrEqual;
-					break;
-				case NOT_EQUAL:
-					d_comp = discord::LobbySearchComparison::NotEqual;
-					break;
-				default:
-					continue;
-			}
-
-			query.Filter(property.utf8(), d_comp, d_cast, s_value.utf8());
-		}
-	}
-
-	_core->LobbyManager().Search(query, [this](discord::Result result) {
-		ERR_FAIL_COND(result != discord::Result::Ok);
-
-		Vector<Variant> vec;
-		int64_t lobby_id;
-		discord::Lobby lobby;
-		int32_t lobby_count;
-		_core->LobbyManager().LobbyCount(&lobby_count);
-
-		for (int32_t i = 0; i < lobby_count; i++) {
-			_core->LobbyManager().GetLobbyId(i, &lobby_id);
-			_core->LobbyManager().GetLobby(lobby_id, &lobby);
-
-			GodotcordLobby gd_lobby;
-			gd_lobby.id = lobby.GetId();
-			gd_lobby.secret = lobby.GetSecret();
-			gd_lobby.max_users = lobby.GetCapacity();
-			gd_lobby.owner_id = lobby.GetOwnerId();
-			_core->LobbyManager().MemberCount(lobby_id, &(gd_lobby.current_users));
-
-			vec.push_back(GodotcordLobby::get_dictionary(&gd_lobby));
-		}
-
-		emit_signal("search_result", vec);
-	});
-}
-
-void Godotcord::get_lobbies(int p_count) {
-	discord::LobbySearchQuery query;
-	_core->LobbyManager().GetSearchQuery(&query);
-
-	_core->LobbyManager().Search(query, [this](discord::Result result) {
-		ERR_FAIL_COND(result != discord::Result::Ok);
-
-		Vector<Variant> vec;
-		int64_t lobby_id;
-		discord::Lobby lobby;
-		int32_t lobby_count;
-		_core->LobbyManager().LobbyCount(&lobby_count);
-
-		for (int32_t i = 0; i < lobby_count; i++) {
-			_core->LobbyManager().GetLobbyId(i, &lobby_id);
-			_core->LobbyManager().GetLobby(lobby_id, &lobby);
-
-			GodotcordLobby gd_lobby;
-			gd_lobby.id = lobby.GetId();
-			gd_lobby.secret = lobby.GetSecret();
-			gd_lobby.max_users = lobby.GetCapacity();
-			gd_lobby.owner_id = lobby.GetOwnerId();
-			_core->LobbyManager().MemberCount(lobby_id, &(gd_lobby.current_users));
-
-			vec.push_back(GodotcordLobby::get_dictionary(&gd_lobby));
-		}
-
-		emit_signal("search_result", vec);
-	});
 }
 
 Array Godotcord::filter_relationships(Object* p_object, StringName p_func_name) {
