+++ godotcord.h
@@ -14,28 +14,6 @@
 class Godotcord : public Object {
     GDCLASS(Godotcord, Object)
 
-public:
-	enum LobbyDistance {
-		LOCAL,
-		DEFAULT,
-		EXTENDED,
-		GLOBAL
-	};
-
-	enum LobbyComparison {
-		LESS_THAN_OR_EQUAL,
-		LESS_THAN,
-		EQUAL,
-		GREATER_THAN,
-		GREATER_THAN_OR_EQUAL,
-		NOT_EQUAL
-	};
-
-	enum LobbyCast {
-		STRING,
-		INT,
-	};
-
 private:
 	discord::Core *_core{};
 	bool init_bool = false;
@@ -58,13 +36,6 @@ public:
 	String get_current_user_discriminator();
 	int64_t get_current_user_id();
 
-	//LOBBY
-	void set_lobby_metadata(int64_t lobby_id, String key, String value);
-	String get_lobby_metadata(int64_t lobby_id, String key);
-
-	void search_lobbies(Variant params, int limit);
-	void get_lobbies(int p_count);
-
 	//RELATIONSHIP
 	Array filter_relationships(Object* p_object, StringName p_func_name);
 	Array get_relationsips();
@@ -81,9 +52,5 @@ public:
 
     Godotcord();
 };
-
-VARIANT_ENUM_CAST(Godotcord::LobbyDistance);
-VARIANT_ENUM_CAST(Godotcord::LobbyCast);
-VARIANT_ENUM_CAST(Godotcord::LobbyComparison);
 
 #endif
